abstracted out bounding box whoops
