+++ chariot-server/src/physics/mod.rs
@@ -1,7 +1,6 @@
-mod collisions;
 pub mod player_entity;
 pub mod trigger_entity;
-mod bounding_box;
+pub mod bounding_box;
 
 #[cfg(test)]
 mod tests;
