+++ chariot-server/src/physics/tests.rs
@@ -6,6 +6,7 @@ use chariot_core::lap_info::LapInformation;
 use chariot_core::player_inputs::EngineStatus;
 use chariot_core::player_inputs::PlayerInputs;
 use chariot_core::player_inputs::RotationStatus;
+use crate::physics::bounding_box::BoundingBox;
 
 use crate::physics::player_entity::PlayerEntity;
 
@@ -28,7 +29,7 @@ fn test_accelerating() {
         mass: 10.0,
 
         size: DVec3::new(10.0, 10.0, 10.0),
-        bounding_box: [[-5.0, 5.0], [-5.0, 5.0], [-5.0, 5.0]],
+        bounding_box: BoundingBox::new(-5.0, 5.0, -5.0, 5.0, -5.0, 5.0),
         physics_changes: Vec::new(),
         lap_info: LapInformation::new(),
     };
@@ -72,7 +73,7 @@ fn test_non_accelerating() {
         mass: 10.0,
 
         size: DVec3::new(10.0, 10.0, 10.0),
-        bounding_box: [[15.0, 25.0], [25.0, 35.0], [35.0, 45.0]],
+        bounding_box: BoundingBox::new(15.0, 25.0, 25.0, 35.0, 35.0, 45.0),
         physics_changes: Vec::new(),
         lap_info: LapInformation::new(),
     };
@@ -114,7 +115,7 @@ fn test_decelerating() {
         mass: 10.0,
 
         size: DVec3::new(10.0, 10.0, 10.0),
-        bounding_box: [[15.0, 25.0], [25.0, 35.0], [35.0, 45.0]],
+        bounding_box: BoundingBox::new(15.0, 25.0, 25.0, 35.0, 35.0, 45.0),
         physics_changes: Vec::new(),
         lap_info: LapInformation::new(),
     };
@@ -159,7 +160,7 @@ fn test_spinning() {
         mass: 10.0,
 
         size: DVec3::new(10.0, 10.0, 10.0),
-        bounding_box: [[15.0, 25.0], [25.0, 35.0], [35.0, 45.0]],
+        bounding_box: BoundingBox::new(15.0, 25.0, 25.0, 35.0, 35.0, 45.0),
         physics_changes: Vec::new(),
         lap_info: LapInformation::new(),
     };
@@ -175,4 +176,147 @@ fn test_spinning() {
         props.angular_velocity,
         GLOBAL_CONFIG.car_spin * GLOBAL_CONFIG.rotation_reduction_coefficient
     );
+}
+
+
+fn get_origin_cube() -> PlayerEntity {
+    return PlayerEntity {
+        player_inputs: PlayerInputs {
+            engine_status: EngineStatus::Neutral,
+            rotation_status: RotationStatus::NotInSpin,
+        },
+
+        entity_location: EntityLocation {
+            position: DVec3::new(0.0, 0.0, 0.0),
+            unit_steer_direction: DVec3::new(1.0, 0.0, 0.0),
+            unit_upward_direction: DVec3::new(0.0, 1.0, 0.0),
+        },
+
+        velocity: DVec3::new(0.0, 0.0, 0.0),
+        angular_velocity: 0.0,
+        mass: 10.0,
+
+        size: DVec3::new(10.0, 10.0, 10.0),
+        bounding_box: BoundingBox::new(-5.0, 5.0, -5.0, 5.0, -5.0, 5.0),
+        physics_changes: Vec::new(),
+        lap_info: LapInformation::new(),
+    };
+}
+
+#[test]
+fn test_collision_with_self() {
+    let origin_cube = get_origin_cube();
+    assert!(origin_cube.bounding_box.is_colliding(&origin_cube.bounding_box));
+}
+
+#[test]
+fn test_engulfed_collision() {
+    let big_origin_cube = get_origin_cube();
+    let mut smol_origin_cube = get_origin_cube();
+    smol_origin_cube.size = DVec3::new(1.0, 1.0, 1.0);
+    smol_origin_cube.update_bounding_box();
+    assert!(big_origin_cube.bounding_box.is_colliding(&smol_origin_cube.bounding_box));
+}
+
+#[test]
+fn test_collision_on_corner() {
+    let origin_cube = get_origin_cube();
+    let mut not_origin_cube = get_origin_cube();
+    not_origin_cube.entity_location.position = DVec3::new(10.0, 10.0, 10.0);
+    not_origin_cube.update_bounding_box();
+    assert!(origin_cube.bounding_box.is_colliding(&not_origin_cube.bounding_box));
+}
+
+#[test]
+fn test_noncollision_on_corner() {
+    let origin_cube = get_origin_cube();
+    let mut not_origin_cube = get_origin_cube();
+    not_origin_cube.entity_location.position = DVec3::new(10.1, 10.1, 10.1);
+    not_origin_cube.update_bounding_box();
+    assert!(!origin_cube.bounding_box.is_colliding(&not_origin_cube.bounding_box));
+}
+
+// we have different logic for the y-direction, might as well test that
+#[test]
+fn test_noncollision_when_above_or_below() {
+    let origin_cube = get_origin_cube();
+    let mut high_cube = get_origin_cube();
+    let mut low_cube = get_origin_cube();
+    high_cube.entity_location.position = DVec3::new(0.0, 20.0, 0.0);
+    low_cube.entity_location.position = DVec3::new(0.0, -20.0, 0.0);
+    high_cube.update_bounding_box();
+    low_cube.update_bounding_box();
+    assert!(!origin_cube.bounding_box.is_colliding(&high_cube.bounding_box));
+    assert!(!origin_cube.bounding_box.is_colliding(&low_cube.bounding_box));
+    assert!(!high_cube.bounding_box.is_colliding(&low_cube.bounding_box)); // just for good measure
+}
+
+#[test]
+fn test_collision_when_above_or_below() {
+    let origin_cube = get_origin_cube();
+    let mut high_cube = get_origin_cube();
+    let mut low_cube = get_origin_cube();
+    high_cube.entity_location.position = DVec3::new(0.0, 8.0, 0.0);
+    low_cube.entity_location.position = DVec3::new(0.0, -8.0, 0.0);
+    high_cube.update_bounding_box();
+    low_cube.update_bounding_box();
+    assert!(origin_cube.bounding_box.is_colliding(&high_cube.bounding_box));
+    assert!(origin_cube.bounding_box.is_colliding(&low_cube.bounding_box));
+    assert!(!high_cube.bounding_box.is_colliding(&low_cube.bounding_box)); // just for good measure
+}
+
+#[test]
+fn test_collision_on_rotated_edges() {
+    // uwu w-wat if i was a cube with edge wength 10 (⁄ ⁄•⁄ω⁄•⁄ ⁄) centewed
+    // at the o-owigin and wotated 45 degwees (>ω^)
+    // a-and u (☆ω☆) wewe a cube with edge wength 10 a-awso wotated 45
+    // degwees (o･ω･o) but c-centewed 10sqwt(2) units away in the
+    // x-diwection Owo a-and we 😳 👉 👈 t-touched edges 🥺
+    let mut owo_cube = get_origin_cube();
+    let mut uwu_cube = get_origin_cube();
+
+    owo_cube.entity_location.unit_steer_direction =
+        DVec3::new(2.0_f64.sqrt() / 2.0, 0.0, 2.0_f64.sqrt() / 2.0);
+    uwu_cube.entity_location.unit_steer_direction =
+        DVec3::new(2.0_f64.sqrt() / 2.0, 0.0, 2.0_f64.sqrt() / 2.0);
+
+    uwu_cube.entity_location.position = DVec3::new(10.0 * 2.0_f64.sqrt() - 0.1, 0.0, 0.0);
+    uwu_cube.update_bounding_box();
+    owo_cube.update_bounding_box();
+    assert!(uwu_cube.bounding_box.is_colliding(&owo_cube.bounding_box));
+}
+
+#[test]
+fn test_collision_on_30_deg_rotated_edges() {
+    // uwu owo yadda yadda
+    let mut owo_cube = get_origin_cube();
+    let mut uwu_cube = get_origin_cube();
+
+    // upper right corner is located at x = 5sqrt(6) / 2, z = 5sqrt(2) / 2
+    owo_cube.entity_location.unit_steer_direction =
+        DVec3::new(1.0 / 2.0, 0.0, 3.0_f64.sqrt() / 2.0);
+    uwu_cube.entity_location.unit_steer_direction =
+        DVec3::new(-1.0 / 2.0, 0.0, 3.0_f64.sqrt() / 2.0);
+
+    uwu_cube.entity_location.position = DVec3::new(5.0 * 6.0_f64.sqrt(), 0.0, 0.0);
+    uwu_cube.update_bounding_box();
+    owo_cube.update_bounding_box();
+    assert!(uwu_cube.bounding_box.is_colliding(&owo_cube.bounding_box));
+}
+
+#[test]
+fn test_3d_bounding_box() {
+    let mut cube = get_origin_cube();
+
+    cube.size = DVec3::new(1.0, 10000.0, 1.0);
+    cube.entity_location.unit_upward_direction =
+        DVec3::new(2.0_f64.sqrt() / 2.0, 2.0_f64.sqrt() / 2.0, 0.0);
+    cube.update_bounding_box();
+
+    let BoundingBox {min_y, max_y, ..} = cube.bounding_box;
+
+    let actual_top = (10_000.0 / 2.0) / (2.0_f64.sqrt());
+    let actual_bottom = (-10_000.0 / 2.0) / (2.0_f64.sqrt());
+    assert!(actual_top * 0.999 < max_y && max_y < actual_top * 1.001);
+    assert!(actual_bottom * 0.999 > min_y && min_y > actual_bottom * 1.001);
 }
